pyproject.toml
[project]
name = "version-2-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
]
